+++ foo_discord_rich/resource.h
@@ -32,6 +32,10 @@
 #define IDC_TEXTBOX_ARTWORK_COMMAND     1094
 #define IDC_CHECK_UPLOAD_ARTWORK        1095
 #define IDC_TEXTBOX_METADB_KEY          1096
+#define IDC_RADIO_TYPE_PLAYING          1097
+#define IDC_RADIO_TYPE_LISTENING        1098
+#define IDC_RADIO_TYPE_WATCHING         1099
+#define IDC_RADIO_TYPE_COMPETING        1100
 #ifndef IDC_STATIC
 #define IDC_STATIC -1
 #endif
@@ -42,7 +46,7 @@
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        123
 #define _APS_NEXT_COMMAND_VALUE         40012
-#define _APS_NEXT_CONTROL_VALUE         1097
+#define _APS_NEXT_CONTROL_VALUE         1101
 #define _APS_NEXT_SYMED_VALUE           101
 #endif
 #endif